requirements.txt
streamlit==1.16.0
langchain-community==0.0.10
sentence-transformers==2.2.0
numpy==1.24.1
faiss-cpu==1.7.4
google-generativeai==0.2.0
pickle5==0.0.11
